requirements.txt
pandas
numpy
yfinance
pytz
TA-Lib
scikit-learn
tensorflow
pyyaml
flask
